Add Problematic Argument to Error Messages
This commit adds functionality to the error handling in all of the
programs. Whenever a bad argument is entered, the error message displays
what the error is, as well as the erroneous argument. This information,
in addition to the usage message, makes it easy to discover where the
problematic syntax occurs.