Add Doug
This commit adds the program Doug, a recreation of du. The only flag is
for help. The path can be specified or uses the current directory. The
output is the size of each subdirectory within the path, as well as the
size of the specified path. Symlinks are ignored, and file sizes are
human readable.